+++ www/Library/admin_widgets/Explorer/res/jquery.ui.Explorer/jquery.ui.Explorer.js
@@ -1,21 +1,28 @@
+/**
+ * Виджет обозревателя объектов
+ * Визуализирует выделение объектов
+ * Query UI widget
+ * Copyright 2012 (C) Boolive
+ */
 (function($) {
 	$.widget("boolive.Explorer", $.boolive.AjaxWidget, {
 
-        _create: function() {
-			$.boolive.AjaxWidget.prototype._create.call(this);
-            var self = this;
-
-            // Выделение объекта
-            $(document).on('after-select-object', function(e, state){
-                self.element.find('.selected').removeClass('selected');
-                if (state.select){
-                    self.element.find('[data-object="'+self.escape(state.select)+'"]').parent().addClass('selected');
-                }
-			});
+        _create: function(){
+            $.boolive.AjaxWidget.prototype._create.call(this);
+            this.after_setState(this.before('getState'), {select: true});
         },
-
-		destroy: function() {
-			$.boolive.AjaxWidget.prototype.destroy.call(this);
-		}
+        /**
+         * Выделение объекта
+         * @param state
+         * @param changes
+         */
+        after_setState: function(state, changes){
+            if ('select' in changes){
+                this.element.find('.selected').removeClass('selected');
+                if (state.select){
+                    this.element.find('[data-object="'+this.escape(state.select)+'"]').addClass('selected');
+                }
+            }
+        }
 	})
 })(jQuery);