+++ www/Library/layouts/Admin/BreadcrumbsMenu/res/jquery.ui.BreadcrumbsMenu/jquery.ui.BreadcrumbsMenu.js
@@ -9,15 +9,30 @@
         _create: function() {
 			var self = this;
 			$.boolive.AjaxWidget.prototype._create.call(this);
+            // Нажатие по пункту
+			this.element.on('click', 'li a', function(e){
+                e.preventDefault();
+                // Вход в объект
+                self.before('setState', [{
+                    object: $(this).attr('href')
+                }]);
+			});
+		},
 
-            $(document).on('after-entry-object', function(e, state, callback){
+        /**
+         * При входе в объект - обновление элементов пути
+         * @param state
+         * @param changes
+         */
+        after_setState: function(state, changes){
+            if ('object' in changes){
                 var uri = state.object;
-                var item = self.element.find('li a[href="'+uri+'"]');
+                var item = this.element.find('li a[href="'+uri+'"]');
                 if (item.size()==0){
                     var path = '/';
                     var names = uri.split('/');
                     var cnt = names.length;
-                    var ul = self.element.find('> ul:first');
+                    var ul = this.element.find('> ul:first');
                     var tab = ul.children('li:first').removeClass('active');
                     ul.empty();
                     ul.append(tab.css('z-index', cnt+1).clone());
@@ -28,22 +43,12 @@
                         ul.append(tab.clone());
                         path += '/';
                     }
-                    self.element.find('li a[href="'+uri+'"]').parent().addClass('active');
+                    this.element.find('li a[href="'+uri+'"]').parent().addClass('active');
                 }else{
-                    self.element.find('li').removeClass('active').removeClass('preactive');
-				    self.element.find('li a[href="'+uri+'"]').parent().addClass('active');
+                    this.element.find('li').removeClass('active').removeClass('preactive');
+                    this.element.find('li a[href="'+uri+'"]').parent().addClass('active');
                 }
-            });
-
-            // Нажатие по пункту
-			self.element.on('click', 'li a', function(e){
-                e.preventDefault();
-				self.element.trigger('before-entry-object', [$(this).attr('href')]);
-			});
-		},
-
-		destroy: function() {
-			$.boolive.wgWidget.prototype.destroy.call(this);
-		}
+            }
+        }
 	})
 })(jQuery);