+++ www/Library/views/Html/Html.php
@@ -8,7 +8,6 @@
 namespace Library\views\Html;
 
 use Boolive\values\Rule,
-    Boolive\commands\Commands,
     Boolive\input\Input,
     Library\views\Widget\Widget;