+++ www/Site.php
@@ -1,12 +1,12 @@
 <?php
 /**
  * Сайт
- * Логика корневого объекта. При исполнеии запускает интерфейс.
+ * Логика корневого объекта. При исполнении запускает интерфейс.
  * @version 1.0
  */
 use Library\views\View\View;
 
-class Site extends \Library\views\View\View
+class Site extends View
 {
     function work()
     {